[FIX] web: Kanban grouped on field with readonly expression
As a reminder, the readonly expression in grouped kanban can enable or
disable the drag and drop feature. If the view is not reaonly but the
field in the arch on which the view is grouped is readonly, the drag
would be disabled.

Before this commit, there was a regression introduced in commit
ba1a5509fa49fd846739252d16083dd8cb334b53 that made the kanban record not
able to evaluate the readonly expression in the fields.

This commit fixes the issue by making the kanban record evaluate the
provided expression. It does not however correctly evaluate the
expression if it contains another field as the drag and drop feature
can be enabled at a view level, not at a per record level. It would then
resolve to not readonly.

Task id 3611116

closes odoo/odoo#148965

X-original-commit: 2b9717400aeb9afbdb73d6b57f46a59992606abe
Signed-off-by: Francois Georis (fge) <fge@odoo.com>
Signed-off-by: Simon Genin (ges@odoo) <ges@odoo.com>